Added all files via upload
This repository contains all files that were used throughout the development process, including prototype images and pseudocode that was used while programming the game. Upon downloading all the files, only the "final project main code.py" file should be run in order to play the game. There is additional information about the game such as controls and objectives in the options menu.